+++ lib/node_probe_web/live/pulse_live.ex
@@ -1,8 +1,6 @@
 defmodule NodeProbeWeb.PulseLive do
   use NodeProbeWeb, :live_view
 
-  alias NodeProbe.Metrics
-
   @impl true
   def mount(_params, _session, socket) do
     if connected?(socket) do